acousticRun	R Documentation

Design network with the provided parameters.

Description

This is the main routine that calls several sub-routines when designing the network. First the input parameters are checked for validity and default values are assigned if input values are missing. Then bathymetry is loaded and the fish distribution generated. Then the goodness grid (sumGrid) is calculated depending on the design parameters (this is the heavy part). When the sumGrid is finished sensors can be placed optimally and stats and figures generated.

Usage

  acousticRun(params, showPlots = FALSE, debug = FALSE,
    opt = FALSE)

Arguments

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

showPlots
If TRUE plots are shown on the screen, if FALSE plots are stored in the img folder.

debug
If enabled, turns on debug printing (console only).

opt
If TRUE use vectorized R commands (faster).

Value

A dictionary of return objects, see RETURN_DESCRIPTIONS.html for more info.