acousticTest	R Documentation

Executes a test run of the program, using default parameters. No additional parameters are necessary.

Description

Executes a test run of the program, using default parameters. No additional parameters are necessary.

Usage

  acousticTest(bias = 1, showPlots = TRUE, debug = FALSE)

Arguments

bias
Choose between bias 1 (fish only), 2 (shadowing only) or 3 (fish and shadowing).

showPlots
If TRUE plots are shown on the screen, if FALSE plots are stored in the img folder.

debug
If enabled, turns on debug printing (console only).

Value

A dictionary of return objects, see RETURN_DESCRIPTOINS.html for more info.