calc.percent.viz	R Documentation

Calculates the percentage of the water column in the surrounding cells that is visible to a sensor placed in the current cell.

Description

Calculates a matrix centered around the current cell (r,c) and containing the percentage of the water column in the surrounding cells that is visible to a sensor placed in the current cell.

Usage

  calc.percent.viz(r, c, rind, cind, bGrid, land,
    sensorDepth, dpflag, params, debug = FALSE)

Arguments

r
Row of the current cell in the bGrid.

c
Column of the current cell in the bGrid.

rind
Row indices of the bGrid to calculate visibility percentage.

cind
Column indices of the bGrid to calculate visibility percentage.

bGrid
A valid bGrid.

land
Matrix containing logicals (TRUE = land cell) indicating wheter a cell in the bGrid is a land cell.

debug
If enabled, turns on debug printing (console only).

sensorDepth
Depth of sensor in current cell.

dpflag
If TRUE depth preference is used meaning that the percentage of visible fish is calculated, if FALSE visible water column is calculated.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

Value

Returns a dictionary with three keys (all vectors): percentVisibility contains the percentage fish/signals visible in the surrounding cells, linearIndex contains the linear indices in the bGrid to which the visibilities pertain, dists contains the distance from the current cell to each of the returned cells as given by linearIndex.