checkForMin	R Documentation

Checks that value is greater than the provided minVal, throwing an error if it isn't.

Description

Checks taht value is greater than minVal, printing an error with the provided name if it isn't.

Usage

  checkForMin(name, value, minVal, stop = TRUE)

Arguments

name
A text name for the value variable (this is what gets printed in the error message.

value
The value to test.

minVal
The minimum value for value.

stop
If TRUE, stops the program when an error occurs.

Value

Nothing.