checkLOS	R Documentation

Returns the percent of the water column visible at a target cell from a starting cell.

Description

If 'depth_off_bottom' and 'depth_off_bottom_sd' are keys in 'params', then the algorithm assumes a normal distribution of fish within the specified zone and will return the integral of the visible zone. Otherwise, it assumes an equal distribution of fish throughout the watercolumn.

Usage

  checkLOS(bGrid, startingCell, targetCell, params,
    debug = FALSE)

Arguments

bGrid
A valid BGrid.

startingCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor's location on the BGrid.

targetCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen tag's location on the BGrid.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

Value

The percent of the watercolumn that is visible (as a double between 0 and 1).