checkStatus	R Documentation

Returns the percent completion of the calculation of the sumgrid for a given job.

Description

Returns the percent completion of the calculation of the sumgrid for a given job.

Usage

  checkStatus(id)

Arguments

id
The id of the job to query. This is always the associated timestamp for the job.

Value

The status of the job as a percentage between zero and one.