conv.1D	R Documentation

Performs the convolution operation in 1D.

Description

Performs the convolution operation in 1D.

Usage

  conv.1D(fun, kern)

Arguments

fun
Vector containing values to perform convolution operation on.

kern
One-dimensional convolution kernel given as a vector preferably with odd-numbered length.

Value

A vector containing the result of the convolution operation with same length as fun.