conv.2D	R Documentation

Performs the convolution operation in 2D.

Description

This is a simple generalization of conv.1D to two dimensions in the sense that it performs two 1D convolutions using two 1D kernels (which can be different). This function is not able to do an actual 2D convolution using a convolution matrix (this functionality is not needed here anyway).

Usage

  conv.2D(mat, kx, ky)

Arguments

mat
Matrix containing values to perform convolution operation on

kx
Convolution kernel in x direction

ky
Convolution kernel in y direction

Value

A matrix containing the result of the convolution operation with same dimensions as mat.