detect	R Documentation

Checks if a tag is detectable by a sensor.

Description

Determines the likelihood of a tag at a given position is detectable by a sensor at a given position, using a specific shapeFunction. This function considers Bathymetry and sensor range.

Usage

  detect(bGrid, sensorPos, tagPos, shapeFcn, params,
    debug = FALSE)

Arguments

bGrid
A valid BGrid.

sensorPos
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor's location on the BGrid.

tagPos
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen tag's location on the BGrid.

shapeFcn
The shape function that describes the attenuation of a sensor. See ShapeFunctions.R for a list of supported functions.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

Value

The percent chance of detection as a double between 0 [no chance of detection] and 1 [guaranteed detection].