fish	R Documentation

Generates a Fish Location Grid (FGrid) for the program to use.

Description

Values in the cells of this grid are expressed as a percentage of the total number of fish on the grid.

Usage

  fish(params, bGrid)

Arguments

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

bGrid
A valid BGrid.

Value

An fGrid of the same dimensions as the provided bGrid.