getArea	R Documentation

Get rows and cols of area.

Description

Defines the "shape" of a sensor's range.

Usage

  getArea(loc, dims, range, debug = FALSE)

Arguments

loc
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor location.

dims
The dimensions of the BGrid. Just call dim() on the BGrid for this.

range
The range of the sensor in bathymetric cells.

debug
If enabled, turns on debug printing (console only).

Value

Returns a dictionary of start/end indexes for rows and columns respectively named : rs,re,cs,ce.