getBathy	R Documentation

Generate a bGrid.

Description

Generates a Bathymetric Grid (BGrid) for the program to use. Able to ingest NetCDF, ArcGIS, and ASC file formats.

Usage

  getBathy(inputFile, inputFileType, startX = 0,
    startY = 0, XDist, YDist, seriesName, debug = FALSE)

Arguments

inputFile
The relative path to the file to open.

inputFileType
The type of file to open. Vailid options are "netcdf", "arcgis", and "asc".

startX
Starting index of the BGrid to take from the bathy file.

startY
Starting index of the BGrid to take from the bathy file.

XDist
The width of your desired BGrid.

YDist
the height of your desired BGrid.

seriesName
If you specified netcdf or arcgis, this is the name of the data series to use.

debug
If enabled, turns on debug printing (console only).

Value

A BGrid based on the parameters given. If an error occurs, a default grid is provided.