getCells	R Documentation

Find cell crossed by a beam.

Description

Returns the cells crossed by a beam from the starting cell to the target cell.

Usage

  getCells(startingCell, targetCell, debug = FALSE)

Arguments

startingCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor's location on the BGrid.

targetCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen tag's location on the BGrid.

debug
If enabled, turns on debug printing (console only).

Value

A data.frame containing x and y indicies of cells crossed by a beam from the starting cell to the target cell.