getCells.opt	R Documentation

Returns the cells crossed by a beam from the starting cell to the target cell.

Description

This is an optimized using vectorization. Returns the cells crossed by a beam from the starting cell to the target cell. Note that the starting cell is omitted from the result set.

Usage

  getCells.opt(startingCell, targetCell, debug = FALSE,
    nr = NULL)

Arguments

startingCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor's location on the BGrid.

targetCell
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen tag's location on the BGrid.

debug
If enabled, turns on debug printing (console only).

nr
Number of rows in bGrid.

Value

If nr != NULL a vector is returned containing the linear indices (not row and col) of cells in the bGrid crossed by a beam from the starting cell to the target cell. If nr == NULL a matrix with a row column and a col column is returned.