getStats	R Documentation

Provides statistical data on detection, given a particular bGrid, fGrid, and sensor arrangement.

Description

This function calculates placements of projected sensors and the value of each sensor (requested and projected) as given in increase in unique recovery rate. Additionally, the acoustic coverage map, unique recovery rate, absolute recovery rate, and sparsity are calculated and returned after placing the requested number of sensors.

Usage

  getStats(params, bGrid, fGrid, sensors, debug = FALSE)

Arguments

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

bGrid
A valid BGrid.

fGrid
A valid FGrid.

sensors
The result of a successful call to sensorFun().

debug
If enabled, turns on debug printing (console only).

Value

A dictionary of statistical values containing the keys: "delta", "sensorMat" "uniqRRs", "acousticCoverage", "absRecoveryRate", "uniqRecoveryRate".