graph	R Documentation

Generates .png files for the visualizations of various grids. You must have write access to the R working directory that the program is executed from. Additionally, ensure that an 'img' folder exists there.

Description

Generates .png files for the visualizations of various grids. You must have write access to the R working directory that the program is executed from. Additionally, ensure that an 'img' folder exists there.

Usage

  graph(result, params, plot.bathy = TRUE)

Arguments

result
A dictionary of return objects, the result of a successfull call to run() or sensorFun().

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

plot.bathy
Specifies whether contour lines for bathymetry should be overlayed in the graphs.

Value

A dictionary containing the filenames of the generated images.