offset	R Documentation

Translates a cartesian point to the center of the gridcell it represents. This is necessary for drawing Line of Sight since we assume sensors are in the center of cells. ex: the cartesian point (3,2) would be converted to (2.5, 1.5), which puts it in the cell located at the third column, second row (aka the cell at (3,2) on a 1-based grid system).

Description

Translates a cartesian point to the center of the gridcell it represents. This is necessary for drawing Line of Sight since we assume sensors are in the center of cells. ex: the cartesian point (3,2) would be converted to (2.5, 1.5), which puts it in the cell located at the third column, second row (aka the cell at (3,2) on a 1-based grid system).

Usage

  offset(point)

Arguments

point
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the point to translate on the BGrid.

Value

A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the translated point.