plot.intersect	R Documentation

Plots line intersects on an existing plot, which must exist.

Description

A horizontal line from zero to rate and a vertial line from zero to n are drawn. This function is useful for highlighting an important value when plotting recovery rate as a function of sensors.

Usage

  plot.intersect(n, rate, col = 1, lty = 1)

Arguments

n
An x value.

rate
A y value.

col
Line color option. See R documentation on the lines() function for more details.

lty
Line type option. See R documentation on the lines() function for more details.

Value

No returned value, lines are drawn on existing plot.