plotGrid	R Documentation

Plots the grid specified by the input type.

Description

In addition to the grid itself sensor locations are also plotted along with numbers indicating the order in which sensors were placed. Furthermore, bathymetry contours can be overlayed using the plot.bathy flag.

Usage

  plotGrid(result, type = "bGrid", main = type, xlab = "",
    ylab = "", plot.bathy = TRUE, plot.sensors = TRUE, ...)

Arguments

result
A dictionary of return objects, the result of a successfull call to run() or sensorFun().

type
Character specifying grid type. Available grids: 'bGrid', 'fGrid', 'sumGrid', or 'acousticCoverage'.

main
Set title of plot.

xlab
Set label of x axis.

ylab
Set label of y axis.

plot.bathy
Specifies whether bathymetry contour lines should be added to plots.

plot.sensors
Specifies if sensors should be added to plot.

...
Additional parameters to image, see ?image.

Value

Nothing.