plotSensors	R Documentation

Adds sensors to current plotGrid plot.

Description

This function assumes a plot created by plotGrid exists.

Usage

  plotSensors(result, circles = TRUE, circlty = 3)

Arguments

result
A dictionary of return objects, the result of a successfull call to run() or sensorFun().

circles
If TRUE circles with radius equal to the detection range are drawn around sensors.

circlty
Line type for circles.

Value

Nothing.