plotUniqueRR	R Documentation

Plot unique recovery rate as a function of placed sensors.

Description

Two-way plot showing cumulative unique recovery rate (top) and gained unique recovery rate (bottom) per sensor as a function of number of sensors (in the order they were placed). This is useful when assessing the value of adding more sensors and when assessing the number of sensors required to obtain a certain recovery rate.

Usage

  plotUniqueRR(result)

Arguments

result
A dictionary of return objects, the result of a successfull call to run() or sensorFun().

Value

Nothing.