sensorFun	R Documentation

Calls functions to generate a 'goodness' grid and choose sensor locations.

Description

Calls functions to generate a 'goodness' grid and choose sensor locations.

Usage

  sensorFun(numSensors, bGrid, fGrid, range, bias, params,
    debug = FALSE, silent = FALSE, save.inter = FALSE)

Arguments

numSensors
The number of sensors the program should place.

bGrid
A valid BGrid.

fGrid
A valid FGrid.

range
The range of the sensor in bathymetric cells.

bias
The goodness algorithm to use, choose 1, 2, or 3. See above for descriptions.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

save.inter
If TRUE intermediary calculations are output as key inter.

silent
If set to TRUE, disables status printing.

Details

Finds a "good" set of sensor placements for a given setup [bGrid, fGrid, params]. Returns a list of locations as grid coordinates, and the calculated sumGrid (goodness grid). Bias value controls the 'goodness' algorithm that gets called. Bias cases: 1. Fish density only. 2. Visibility due to Bathy only. 3. Detectable fish accounting for bathy.

Value

A dictionary of return objects, see RETURN_DESCRIPTIONS.html for more info.