sensorFun.suppressHelper	R Documentation

Performs suppression on the sumGrid/fGrid depending upon the suppressionFcn specified.

Description

The suppressionFcn 'detection.functino.exact' indicates that the fGrid must be suppressed, then the entire sumGrid recalculated based upon the new fGrid. Any other suppressionFcn simply calls suppress.opt, which just suppresses the sumGrid, avoiding the need to recalculate the whole sumGrid (a very expensive operation). The latter approach can be considered a 'fast approximation', while the former an 'exact count'.

Usage

  sensorFun.suppressHelper(loc, grids, range, bias, params,
    debug = FALSE)

Arguments

loc
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor location.

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

range
The range of the sensor in bathymetric cells.

bias
The goodness algorithm to use, choose 1, 2, or 3. See above for descriptions.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

Value

Returns the grids parameter, with an updated FGrid.