shape.gauss	R Documentation

Detection function with the shape of a half Gaussian distribution peaking at distance zero, and declining with increasing distances.

Description

Detection function with the shape of a half Gaussian distribution peaking at distance zero, and declining with increasing distances.

Usage

  shape.gauss(dist, params)

Arguments

dist
The distance from the tag to the reciever.

params
A dictionary containing the keys 'sd' and 'peak', where sd defines a standard deviation, and peak defiens the peak value of the curve (maximum value).

Value

The probability of detecting a tag at a given dist with the given parameters.