simulateBGrid	R Documentation

Creates a default BGrid to use if real data could not be loaded.

Description

Creates a default BGrid to use if real data could not be loaded.

Usage

  simulateBGrid(XDist, YDist)

Arguments

XDist
The width of your desired BGrid.

YDist
the height of your desired BGrid.

Value

A default BGrid of dimensions XDist by YDist.