sub2ind	R Documentation

Converts a row, col index to a linear index within a matrix.

Description

Converts a row, col index to a linear index within a matrix.

Usage

  sub2ind(row, col, dims)

Arguments

row
The row index.

col
The col index.

dims
The dimensions of the BGrid. Just call dim() on the parent matrix for this.

Value

The translated linear index.