sumGrid.sumBathy	R Documentation

Sums the result of calling the detect() function on each cell within range of a target cell for each cell in the given grid.

Description

Sums the result of calling the detect() function on each cell within range of a target cell for each cell in the given grid.

Usage

  sumGrid.sumBathy(grids, range, shapeFcn = "shape.t",
    params, debug = FALSE)

Arguments

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

range
The range of the sensor in bathymetric cells.

shapeFcn
The shape function that describes the attenuation of a sensor. See ShapeFunctions.R for a list of supported functions.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

Value

Returns the grids parameter, with an updated sumGrid.