sumGrid.sumBathy.opt	R Documentation

Calculates the sumGrid when a line of sight bias is chosen (bias 2 or 3).

Description

Loops through all cells where sensor placement is valid (where sensor would be below surface) and calculates goodness. If bias is 2 only bathymetry (line of sight) is used to calculate goodness, whereas if bias is 3 both bathymetry and fish distribution (fGrid) are used. This function uses vectorized calculations.

Usage

  sumGrid.sumBathy.opt(grids, params, debug = FALSE,
    silent = FALSE)

Arguments

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

silent
If set to TRUE, turns off status printing.

Value

Returns the grids parameter, with an updated sumGrid.