sumGrid.sumSimple	R Documentation

Calculates sumGrid for bias 1.

Description

Simply sums the values within range of a cell, for each cell in the given grid.

Usage

  sumGrid.sumSimple(grids, key, range, debug = FALSE)

Arguments

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

key
A key to the dictionary provided in the 'grids' parameter specifying which grid should be summed.

range
The range of the sensor in bathymetric cells.

debug
If enabled, turns on debug printing (console only).

Value

Returns the grids parameter, with an updated sumGrid.