sumGrid.sumSimple.opt	R Documentation

Simply sums the values within range of a cell, for each cell in the given grid.

Description

This is a speed optimized version, which uses the convolution operation (which mainly gains it speed from using FFT) as an alternative to running through tedious R for loops.

Usage

  sumGrid.sumSimple.opt(grids, key, params, debug = FALSE,
    silent = FALSE)

Arguments

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

key
A key to the dictionary provided in the 'grids' parameter specifying which grid should be summed.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

silent
If set to TRUE, disables status printing.

Value

Returns the grids parameter, with an updated sumGrid.