suppress.opt	R Documentation

Suppresses the values of cells around a sensor using a specified suppressionFunction.

Description

This is an optimized version, which uses vectorization.

Usage

  suppress.opt(sumGrid, dims, loc, params, bGrid,
    debug = FALSE)

Arguments

sumGrid
A valid SumGrid.

dims
The dimensions of the BGrid. Just call dim() on the BGrid for this.

loc
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor location.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

bGrid
valid BGrid.

debug
If enabled, turns on debug printing (console only).

Value

Returns a suppressed sumGrid.