suppression.scale	R Documentation

Returns a dynamic value based on distance from a chosen sensor. The returned value should be multiplied by the value to be scaled.

Description

Returns a dynamic value based on distance from a chosen sensor. The returned value should be multiplied by the value to be scaled.

Usage

  suppression.scale(dist, suppressionRange,
    minsuppressionValue, maxsuppressionValue, params,
    debug = FALSE)

Arguments

dist
The distance between two cells on a grid.

suppressionRange
How far out to apply suppression penalties, in bathymetric cells.

minsuppressionValue
The minimum allowable value to return.

maxsuppressionValue
The maximum allowable value to return (also the return value for suppression.static()).

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

Value

Returns The value given in maxSuppression Value.