test	R Documentation

Executes a test run of the program, using default parameters. No additional parameters are necessary.

Description

Executes a test run of the program, using default parameters. No additional parameters are necessary.

Usage

  test(debug = FALSE, opt = FALSE)

Arguments

debug
If enabled, turns on debug printing (console only).

opt
Tells the program to use vectorized R commands.

Value

A dictionary of return objects, see RETURN_DESCRIPTOINS.html for more info.