updateFGrid	R Documentation

Updates the FGrid after each sensor is placed to reflect which areas that are already covered by sensors.

Description

When a sensor is placed the FGrid must be updated to reflect where unique signals are emitted. This is done by calculating the coverage of the sensor given by loc and then downweighing the FGrid using this coverage such that locations that are well covered by the sensor at loc is downweighed more that poorly covered locations.

Usage

  updateFGrid(loc, grids, params, debug = FALSE,
    opt = FALSE)

Arguments

loc
A dictionary containing the keys 'r' and 'c', which hold the row and column indicies of the chosen sensor location.

grids
A dictionary containing the keys 'bGrid', 'fGrid', and 'sumGrid', which hold a valid BGrid, FGrid and SumGrid.

params
A dictionary of parameters, see PARAMETER_DESCRIPTIONS.html for more info.

debug
If enabled, turns on debug printing (console only).

opt
Tells the program to use vectorized R commands.

Value

Returns the grids parameter, with an updated FGrid.