verticalHabitat	R Documentation

A vectorized helper function that indicates if a cell has a depth between mindepth and maxdepth.

Description

Returns boolean grid. Cells between the given depths are set to TRUE, and all other cells are set to FALSE.

Arguments

mindepth
The shallowest depth a fish will visit.

maxdepth
The deepest depth a fish will visit.

bGrid
A valid BGrid.

Value

A boolean grid with values set to TRUE if the cell is between the depth restrictions, and FALSE otherwise.