writeFiles	R Documentation

Adds txt and zip file paths to the filenames attribute, then writes a json file.

Description

Writes the txt and json dumps of raw data used in the job, png files of the graphs generated by the job, and optionally a zip file containing all of the above. All filenames include a the time value so that associated files are recognizable.

Usage

  writeFiles(filenames, result, path, time, zip = TRUE)

Arguments

result
A dictionary of return objects, the result of a successfull call to run() or sensorFun().

filenames
A dictionary containing file paths; keys are: bGrid, fGrid, sumGrid, acousticCoverage, and recoveryRates.

path
A path to prepend to all the filenames.

time
The timestamp label to include in all filenames.

zip
If true, writes a zip file containing the txt dump, and images of a run.

Value

A dictionary containing the filenames of the generated images.